Ajustando Controllers e Banco

+++ Liga-Tabajara/Controllers/JogadoresController.cs
@@ -1,10 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Data;
 using System.Data.Entity;
 using System.Linq;
 using System.Net;
-using System.Web;
 using System.Web.Mvc;
 using Liga_Tabajara.Data;
 
@@ -12,12 +9,30 @@ namespace Liga_Tabajara.Controllers
 {
     public class JogadoresController : Controller
     {
-        private LigaContext db = new LigaContext();
+        private readonly LigaContext db = new LigaContext();
 
         // GET: Jogadores
-        public ActionResult Index()
+        public ActionResult Index(string searchName, Posicao? posicao, PePreferido? pePreferido)
         {
-            var jogadores = db.Jogadores.Include(j => j.Time);
+            var jogadores = db.Jogadores.Include(j => j.Time).AsQueryable();
+
+            // Filtros
+            if (!string.IsNullOrWhiteSpace(searchName))
+                jogadores = jogadores.Where(j => j.Nome.Contains(searchName));
+            if (posicao.HasValue)
+                jogadores = jogadores.Where(j => j.Posicao == posicao.Value);
+            if (pePreferido.HasValue)
+                jogadores = jogadores.Where(j => j.PePreferido == pePreferido.Value);
+
+            // Dropdowns para View
+            ViewBag.Posicoes = Enum.GetValues(typeof(Posicao))
+                                   .Cast<Posicao>()
+                                   .Select(p => new SelectListItem { Text = p.ToString(), Value = p.ToString(), Selected = (p == posicao) });
+            ViewBag.Pes = Enum.GetValues(typeof(PePreferido))
+                                .Cast<PePreferido>()
+                                .Select(p => new SelectListItem { Text = p.ToString(), Value = p.ToString(), Selected = (p == pePreferido) });
+            ViewBag.SearchName = searchName;
+
             return View(jogadores.ToList());
         }
 
@@ -25,14 +40,12 @@ namespace Liga_Tabajara.Controllers
         public ActionResult Details(int? id)
         {
             if (id == null)
-            {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Jogador jogador = db.Jogadores.Find(id);
+
+            var jogador = db.Jogadores.Find(id);
             if (jogador == null)
-            {
                 return HttpNotFound();
-            }
+
             return View(jogador);
         }
 
@@ -44,68 +57,84 @@ namespace Liga_Tabajara.Controllers
         }
 
         // POST: Jogadores/Create
-        // Para proteger-se contra ataques de excesso de postagem, ative as propriedades específicas às quais deseja se associar. 
-        // Para obter mais detalhes, confira https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "Id,Nome,DataNascimento,Nacionalidade,Posicao,NumeroCamisa,Altura,Peso,PePreferido,TimeId")] Jogador jogador)
         {
-            if (ModelState.IsValid)
+            if (!ModelState.IsValid)
             {
-                db.Jogadores.Add(jogador);
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                ViewBag.TimeId = new SelectList(db.Times, "Id", "Nome", jogador.TimeId);
+                return View(jogador);
             }
 
-            ViewBag.TimeId = new SelectList(db.Times, "Id", "Nome", jogador.TimeId);
-            return View(jogador);
+            if (!JogadorValido(jogador))
+            {
+                ModelState.AddModelError("",
+                    "Jogador inválido: nome não vazio, data de nascimento menor que hoje, camisa única no time e enums selecionados.");
+
+                ViewBag.TimeId = new SelectList(db.Times, "Id", "Nome", jogador.TimeId);
+                return View(jogador);
+            }
+
+            db.Jogadores.Add(jogador);
+            db.SaveChanges();
+            return RedirectToAction("Index");
         }
 
         // GET: Jogadores/Edit/5
         public ActionResult Edit(int? id)
         {
             if (id == null)
-            {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Jogador jogador = db.Jogadores.Find(id);
+
+            var jogador = db.Jogadores.Find(id);
             if (jogador == null)
-            {
                 return HttpNotFound();
-            }
+
             ViewBag.TimeId = new SelectList(db.Times, "Id", "Nome", jogador.TimeId);
             return View(jogador);
         }
 
         // POST: Jogadores/Edit/5
-        // Para proteger-se contra ataques de excesso de postagem, ative as propriedades específicas às quais deseja se associar. 
-        // Para obter mais detalhes, confira https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "Id,Nome,DataNascimento,Nacionalidade,Posicao,NumeroCamisa,Altura,Peso,PePreferido,TimeId")] Jogador jogador)
         {
-            if (ModelState.IsValid)
+            if (!ModelState.IsValid)
             {
-                db.Entry(jogador).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                ViewBag.TimeId = new SelectList(db.Times, "Id", "Nome", jogador.TimeId);
+                return View(jogador);
             }
-            ViewBag.TimeId = new SelectList(db.Times, "Id", "Nome", jogador.TimeId);
-            return View(jogador);
+
+            if (!JogadorValido(jogador))
+            {
+                ModelState.AddModelError("",
+                    "Jogador inválido: nome não vazio, data de nascimento menor que hoje, camisa única no time e enums selecionados.");
+
+                ViewBag.TimeId = new SelectList(db.Times, "Id", "Nome", jogador.TimeId);
+                return View(jogador);
+            }
+
+            db.Entry(jogador).State = EntityState.Modified;
+            db.SaveChanges();
+            return RedirectToAction("Index");
         }
 
         // GET: Jogadores/Delete/5
         public ActionResult Delete(int? id)
         {
             if (id == null)
-            {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Jogador jogador = db.Jogadores.Find(id);
+
+            var jogador = db.Jogadores.Find(id);
             if (jogador == null)
-            {
                 return HttpNotFound();
-            }
+
+            // Bloqueia exclusão se houver gols registrados
+            bool temGols = db.Gols.Any(g => g.JogadorId == jogador.Id);
+            if (temGols)
+                ViewBag.Erro = "Não é possível excluir jogador com gols registrados.";
+
             return View(jogador);
         }
 
@@ -114,7 +143,15 @@ namespace Liga_Tabajara.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Jogador jogador = db.Jogadores.Find(id);
+            var jogador = db.Jogadores.Find(id);
+            bool temGols = db.Gols.Any(g => g.JogadorId == jogador.Id);
+
+            if (temGols)
+            {
+                ModelState.AddModelError("", "Não é possível excluir jogador com gols registrados.");
+                return View(jogador);
+            }
+
             db.Jogadores.Remove(jogador);
             db.SaveChanges();
             return RedirectToAction("Index");
@@ -122,11 +159,28 @@ namespace Liga_Tabajara.Controllers
 
         protected override void Dispose(bool disposing)
         {
-            if (disposing)
-            {
-                db.Dispose();
-            }
+            if (disposing) db.Dispose();
             base.Dispose(disposing);
         }
+
+        #region Métodos Privados de Validação
+        private bool JogadorValido(Jogador jogador)
+        {
+            // Nome
+            if (string.IsNullOrWhiteSpace(jogador.Nome))
+                return false;
+            // Data de nascimento
+            if (jogador.DataNascimento >= DateTime.Today)
+                return false;
+            // Numeração única por time
+            var qtMesmoNumero = db.Jogadores
+                .Count(j => j.TimeId == jogador.TimeId && j.NumeroCamisa == jogador.NumeroCamisa && j.Id != jogador.Id);
+            if (qtMesmoNumero > 0)
+                return false;
+            // Enums preenchidos (não dependem de null pois são tipos valor)
+            return Enum.IsDefined(typeof(Posicao), jogador.Posicao)
+                && Enum.IsDefined(typeof(PePreferido), jogador.PePreferido);
+        }
+        #endregion
     }
 }
